+++ lib/crud_components/fields/enum_field.rb
@@ -10,7 +10,7 @@ module CrudComponents
       def derived_sortable? = true
       def derived_filter_control = :select
       def default_editable? = true
-      def form_control = :enum
+      def default_form_control = :enum
 
       def form_choices
         enum_keys.map { |key| [human_value(key), key] }